+++ CHANGELOG.md
@@ -9,6 +9,12 @@ format. This project adheres to [Semantic Versioning](https://semver.org/spec/v2
 
 - Added support for dynamic custom casters
 - Added `LaravelDumper` facade
+- Added custom `Properties` collection for easier manipulation of dumped properties
+
+### Changed
+
+- Changed `Caster` interface to use `Properties` collection
+- Updated all casters to use new `Properties` collection
 
 ## [0.2.0]