+++ CHANGELOG.md
@@ -5,6 +5,8 @@ format. This project adheres to [Semantic Versioning](https://semver.org/spec/v2
 
 ## [Unreleased]
 
+## [0.3.0]
+
 ### Added
 
 - Added support for dynamic custom casters